Add kdtree_match
